% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{package-ptf}
\alias{package-ptf}
\alias{ptf-package}
\alias{package-ptf-package}
\title{probit tensor factorization}
\description{
\tabular{ll}{
Package: \tab ptf \cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2020-09-30\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}
}
\author{
Ye Liu
Maintainer: Ye Liu \email{yeliu.nikki@gmail.com}
}
\keyword{factorization}
\keyword{model}
\keyword{probit}
\keyword{tensor}
