% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rescal.R
\name{Rescal}
\alias{Rescal}
\title{RESCAL Model}
\usage{
Rescal(X, k, n, r, max_iter = 500)
}
\arguments{
\item{X}{response data, which is a three-way array of size n by n by k}

\item{k}{number of relations}

\item{n}{number of entities}

\item{r}{decomposition rank}

\item{max_iter}{max number of iterations}
}
\value{
fitted parameters
}
\description{
RESCAL Model
}
\references{
Nickel, Maximilian, Volker Tresp, and Hans-Peter Kriegel. "A three-way model for collective learning on multi-relational data.", Icml. 2011.
}
